% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_interface.R
\name{convert_q_date}
\alias{convert_q_date}
\title{Convert date if it's quarterly, rather than monthly.
this changes the x axis}
\usage{
convert_q_date(obj)
}
\arguments{
\item{obj}{List containing data, metadata, and update info.}
}
\value{
The modified list.
}
\description{
Convert date if it's quarterly, rather than monthly.
this changes the x axis
}
