% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_interface.R
\name{get_chapters}
\alias{get_chapters}
\title{Get distinct chapter names.}
\usage{
get_chapters()
}
\value{
A data frame containing distinct chapter names in the \code{sol_chapter} column.
Returns an empty data frame if an error occurs.
}
\description{
Retrieves a list of distinct \code{sol_chapter} values from the \code{mtd} table.
}
\examples{
chapters <- get_chapters()
}
