% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_interface.R
\name{get_charts}
\alias{get_charts}
\title{Get charts for a given chapter and subsection.}
\usage{
get_charts(chpt, sub)
}
\arguments{
\item{chpt}{The chapter name.}

\item{sub}{The subsection name.}
}
\value{
A named character vector where the names are chart titles and the
values are dataset names. Returns an empty character vector if an error occurs.
}
\description{
Retrieves a list of dataset names and their corresponding chart titles for a
specific \code{sol_chapter} and \code{sol_subsection} from the \code{mtd} table.
}
\examples{
charts <- get_charts("Introduction", "Overview")
}
