% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_interface.R
\name{get_charts_list}
\alias{get_charts_list}
\title{Get a list of available charts.}
\usage{
get_charts_list()
}
\value{
A named character vector where the names are chart titles and the
values are dataset names. Returns an empty character vector if an error occurs.
}
\description{
Retrieves a list of dataset names and their corresponding chart titles from the
\code{mtd} table in the SQLite database.
}
\examples{
charts <- get_charts_list()
names(charts) # Chart titles
charts        # Dataset names
}
