% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_interface.R
\name{get_data_for_chart}
\alias{get_data_for_chart}
\title{Get data for the chart from the SQLite database.}
\usage{
get_data_for_chart(dtst, sqlite_dpth, chart = NULL)
}
\arguments{
\item{dtst}{Dataset name.}

\item{sqlite_dpth}{Path to the SQLite database.}

\item{chart}{Optional filter chart name. (look to get rid of this)}
}
\value{
A list containing data, metadata, and update information.
}
\description{
Get data for the chart from the SQLite database.
}
