% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chart_data.R
\name{get_obs_chart}
\alias{get_obs_chart}
\title{Get an Observable chart.}
\usage{
get_obs_chart(dtst)
}
\arguments{
\item{dtst}{The dataset name.}
}
\value{
A list containing processed data, chart options, metadata,
the Observable chart, and update information.
}
\description{
Retrieves data from SQLite, processes it, and generates an Observable chart.
}
