% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chart_data.R
\name{orderctg}
\alias{orderctg}
\title{Order data by category values.}
\usage{
orderctg(df, ctgfact)
}
\arguments{
\item{df}{The data frame.}

\item{ctgfact}{Factor levels for the b column.}
}
\value{
The ordered data frame.
}
\description{
Order data by category values.
}
