% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chart_data.R
\name{orderxd}
\alias{orderxd}
\title{Order data by x-axis values.}
\usage{
orderxd(df, xfact)
}
\arguments{
\item{df}{The data frame.}

\item{xfact}{Factor levels for the xd column.}
}
\value{
The ordered data frame.
}
\description{
Order data by x-axis values.
}
