% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_interface.R
\name{refresh_analysis}
\alias{refresh_analysis}
\title{Refresh the analysis table.}
\usage{
refresh_analysis()
}
\value{
None. This function is called for its side effect.
}
\description{
Reads analysis data from a CSV file (\code{data/analysis.csv}), converts the \code{md}
column to HTML using \code{md_convert}, and writes it to the \code{analysis} table in
the SQLite database, overwriting the existing table.
}
\examples{
refresh_analysis()
}
