% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_interface.R
\name{refresh_meta}
\alias{refresh_meta}
\title{Refresh the metadata table.}
\usage{
refresh_meta()
}
\value{
None. This function is called for its side effect.
}
\description{
Reads metadata from a CSV file (\code{data/meta.csv}) and writes it to the \code{mtd}
table in the SQLite database, overwriting the existing table.
}
\examples{
refresh_meta()
}
